+++ lib/inc/lqr_steer_control.h
@@ -13,15 +13,15 @@ class lqr_steer_control : public path_tracker
 {
 public:
     lqr_steer_control();
-    lqr_steer_control(const double max_steer_angle, const double max_speed, const double wheel_base, const double center_to_gps_distance);
+    lqr_steer_control(const float max_steer_angle, const float max_speed, const float wheel_base, const float center_to_gps_distance);
     ~lqr_steer_control();
 
 public:
-    virtual double steering_control(pt_control_state_t state, path_point_t target_point);
-    virtual double velocity_control(pt_control_state_t state, path_point_t target_point);
+    virtual float steering_control(pt_control_state_t state, path_point_t target_point);
+    virtual float velocity_control(pt_control_state_t state, path_point_t target_point);
 
-    virtual void set_gain(int gain_index, double* gain_value);
-    virtual void get_gain(int gain_index, double* gain_value);
+    virtual void set_gain(int gain_index, float* gain_value);
+    virtual void get_gain(int gain_index, float* gain_value);
 
 private:
     ModelMatrix dlqr(ModelMatrix A, ModelMatrix B, ModelMatrix Q, ModelMatrix R);
