+++ lib/inc/pid_steer_control.h
@@ -13,29 +13,29 @@ class pid_steer_control : public path_tracker
 {
 public:
     pid_steer_control();
-    pid_steer_control(const double max_steer_angle, const double max_speed, const double wheel_base, const double center_to_gps_distance);
+    pid_steer_control(const float max_steer_angle, const float max_speed, const float wheel_base, const float center_to_gps_distance);
     ~pid_steer_control();
 
 public:
-    virtual void set_gain(int gain_index, double* gain_value);
-    virtual void get_gain(int gain_index, double* gain_value);
+    virtual void set_gain(int gain_index, float* gain_value);
+    virtual void get_gain(int gain_index, float* gain_value);
 
 private:
-    virtual double steering_control(pt_control_state_t state, path_point_t target_point);
-    virtual double velocity_control(pt_control_state_t state, path_point_t target_point);
+    virtual float steering_control(pt_control_state_t state, path_point_t target_point);
+    virtual float velocity_control(pt_control_state_t state, path_point_t target_point);
 
 private:
     pid_controller path_accel_pid;
     pid_controller path_distance_pid;
 
-    double steer_kp;
-    double steer_ki;
-    double steer_kd;
-    double steer_pre_e;
+    float steer_kp;
+    float steer_ki;
+    float steer_kd;
+    float steer_pre_e;
 
-    double adapted_pid_distance_threshold;
-    double adapted_pid_distance_gain;
-    double adapted_pid_yaw_gain;
+    float adapted_pid_distance_threshold;
+    float adapted_pid_distance_gain;
+    float adapted_pid_yaw_gain;
 
 public:
 };
